move extractor to it's own file
